Added new functionality
Added the ability to change the blank symbol (what erase will write to the tape) and change all of one symbol on the tape to another symbol. Both are untested and should be used with caution as long as the UNTESTED note in the docstring is there.

Also added docstrings to both functions.